Changed the way how GUIDs are generated
When creating a new package in newer versions of Umbraco, GUIDs are not really GUIDs. Umbraco instead uses the name of the file as the value for the GUID. If multiple files have the same name, any additional files will be named like "{guid}_{name}". This commit tries to replicate that.